# :racehorse: use `EntryPart<&[u8]>::split` to avoid copy for performance  improvement