Before
```
$ cargo +nightly bench split
    Finished `bench` profile [optimized] target(s) in 0.64s
     Running unittests src/lib.rs (target/release/deps/libpna-7c13e9ba111e4488)
...
test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 2 filtered out; finished in 0.00s

     Running benches/split.rs (target/release/deps/split-5b7839c129085926)

running 2 tests
test create_with_split ... bench:  16,709,934.20 ns/iter (+/- 11,876,727.33)
test split             ... bench:  10,346,815.10 ns/iter (+/- 10,597,196.51)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 1 filtered out; finished in 8.38s
```

After
```
$ cargo +nightly bench split
    Finished `bench` profile [optimized] target(s) in 0.31s
     Running unittests src/lib.rs (target/release/deps/libpna-7c13e9ba111e4488)
...
test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 2 filtered out; finished in 0.00s

     Running benches/split.rs (target/release/deps/split-5b7839c129085926)

running 2 tests
test create_with_split ... bench:  10,380,744.30 ns/iter (+/- 3,855,647.82)
test split             ... bench:  11,052,741.70 ns/iter (+/- 7,957,370.03)
```